+++ src/main/java/dao/GameDao.java
@@ -1,4 +1,4 @@
-package repository;
+package dao;
 
 import domain.Game;
 import domain.board.TableSetting;
@@ -10,12 +10,11 @@ import java.sql.SQLException;
 import java.sql.Statement;
 import java.util.Optional;
 
-public class GameRepositoryImpl implements GameRepository {
+public class GameDao {
     private static final String FAILED_SAVE_GAME_DATA = "[ERROR] DB에 게임 데이터를 저장하는 도중, 오류가 발생했습니다.";
     private static final String FAILED_FIND_LATEST_GAME_DATA = "[ERROR] DB에서 최근 게임 데이터를 조회하는 도중, 오류가 발생했습니다.";
     private static final String FAILED_UPDATE_GAME_FINISHED = "[ERROR] DB에서 게임 종료 처리 도중, 오류가 발생했습니다.";
 
-    @Override
     public Optional<Long> save(Connection connection, Game game) {
         String sql = "INSERT INTO game(created_at, cho_table_setting, han_table_setting) VALUES (?, ?, ?)";
 
@@ -38,7 +37,6 @@ public class GameRepositoryImpl implements GameRepository {
         return Optional.empty();
     }
 
-    @Override
     public Optional<Game> findLatest(Connection connection) {
         String sql = "SELECT  * FROM game WHERE is_finished = FALSE ORDER BY game_id DESC LIMIT 1";
 
@@ -58,7 +56,6 @@ public class GameRepositoryImpl implements GameRepository {
         return Optional.empty();
     }
 
-    @Override
     public void finishedGame(Connection connection, Long gameId) {
         String sql = "UPDATE game SET is_finished = TRUE WHERE game_id = ?";
