--- src/main/java/repository/GameRepository.java
@@ -1,13 +1,0 @@
-package repository;
-
-import domain.Game;
-import java.sql.Connection;
-import java.util.Optional;
-
-public interface GameRepository {
-    Optional<Long> save(Connection connection, Game game);
-
-    Optional<Game> findLatest(Connection connection);
-
-    void finishedGame(Connection connection, Long gameId);
-}
